+++ fasttext/interface.h
@@ -47,7 +47,7 @@ class FastTextModel {
                 std::shared_ptr<Matrix> output);
         void setModel(std::shared_ptr<Model> model);
 };
-typedef void (*callbackfunc)(int progress, void *user_data);
+typedef void (*callbackfunc)(double progress, void *user_data);
 void trainWrapper(int argc, char **argv, int silent, callbackfunc callback_run, void* callback);
 void loadModelWrapper(std::string filename, FastTextModel& model);
